Big Error Checking Update (read)
This update should hopefully solve almost all of the error checking. I made some small changes like having all inputs read as Strings so that we can check the validity of the input. NumberTickets, for example used to read the number as an int which makes sense but if the user entered a letter the code would crash. Now we can safely check all inputs to make sure they fit within the constraints that they are supposed to. Agent privileges have also been implemented so that they can not manipulate tickets over the limits set in the assignment.