Small Usability Update
I just made some quick edits to the code to improve the usability of the program, such as actually confirming that we have successfully made a transaction because before it would print something very close to the error message we print which may have caused issues. Now it is certain you successfully made the transaction.